Removed portfolio Value component and moved functionality into container. Moved the number to center of pie chart
