+++ client/src/containers/SideNav.js
@@ -6,12 +6,11 @@ import _ from 'lodash';
 import ReactInterval from 'react-interval';
 import { Row, Col, Preloader, Table, Collapsible, CollapsibleItem, Card, CardTitle, Collection, CollectionItem } from 'react-materialize';
 import { PieChart, Sector, Pie, Cell, ResponsiveContainer, Legend, Tooltip } from 'recharts';
+import Divider from 'material-ui/Divider';
 
 import { TYPE } from '../actions/types';
 import Tracker from './Tracker';
-import PortfolioValue from '../components/PortfolioValue';
 import { addTicker, loadTickerList, loadTickerPrices, removeTicker, loadChartData, updateQuantity, selectChart } from '../actions/index';
-
 // http://recharts.org/#/en-US/guide/getting-started
 // For dynamic pie chart documentation
 
@@ -120,7 +119,7 @@ class SideNav extends Component {
                      <Col s={2}>
                         <div className="quantity-text">Quantity</div>
                      </Col>
-                  </Row>
+                  </Row><hr className="margin: 0px"></hr>
                {this.props.tickerList.map(ticker =>
                   this.renderTracker(ticker, renderType)
                )}
@@ -157,11 +156,36 @@ class SideNav extends Component {
       return pieChartData;
    }
 
+   calculateValue() {
+      const { priceList, tickerList } = this.props;
+
+      if(priceList && tickerList) {
+         let currentValue = 0;
+
+         for (let i = 0; i < tickerList.length; i++) {
+            if (tickerList[i]) {
+               const { name, type, quantity } = tickerList[i];
+
+               if (priceList[type] && priceList[type][name]) {
+                  currentValue += priceList[type][name] * quantity;
+               }
+            }
+         }
+         const numberWithCommas = (x) => {
+            return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+         }
+         currentValue = Number(currentValue).toFixed(2);
+
+         return currentValue;
+      }
+      return 0;
+   }
+
    render() {
       const refreshRateSeconds = 15;
       const timeout = refreshRateSeconds * 1000;
 
-      const COLORS = ['#8884d8'];
+      const COLORS = ['#ED5281'];
 
       const data = this.getPieChartData();
 
@@ -173,26 +197,18 @@ class SideNav extends Component {
                callback={this.props.loadTickerPrices}
             />
             <ul id="nav-mobile" className="side-nav fixed z-depth-8">
-               <Card
-                  className="navbar-img"
-                  header={
-                     <CardTitle image={require('../img/a.jpg')}>
-                        <PieChart width={300} height={300} onMouseEnter={this.onPieEnter}>
-                           <Pie data={data} cx={155} cy={200} innerRadius={70} outerRadius={80} fill="#8884d8" paddingAngle={6}>
-                              {data.map((entry, index) => (
-                                 <Cell fill={COLORS[index % COLORS.length]} />
-                              ))}
-                           </Pie>
-                           <Tooltip />
-                        </PieChart>
-                     </CardTitle>
-                  }
-               >
-                  <PortfolioValue
-                     tickerList={this.props.tickerList}
-                     priceList={this.props.priceList}
-                  />
-               </Card>
+              <PieChart width={280} height={280} onMouseEnter={this.onPieEnter}>
+                <text className="portfolio-value" x={175} y={175} dy={8} textAnchor="middle" fill="#FFFFFF">
+                  ${this.calculateValue()}
+                </text>
+                 <Pie data={data} cx={170} cy={170} innerRadius={70} outerRadius={80} fill="#8884d8" paddingAngle={6}>
+                    {data.map((entry, index) => (
+                       <Cell fill={COLORS[index % COLORS.length]} />
+                    ))}
+                 </Pie>
+                 <Tooltip />
+               </PieChart>
+               <Divider />
                <Collapsible className="ticker-collasp">
                   <CollapsibleItem
                      id="collapsible-header"
