+++ package.json
@@ -42,10 +42,14 @@
       "passport-local-mongoose": "^4.4.0",
       "popper.js": "^1.12.9",
       "react": "^16.2.0",
+      "react-animated-number": "^0.4.3",
       "react-chartjs-2": "^2.6.4",
       "react-dom": "^16.2.0",
       "react-interval": "^2.0.2",
       "react-materialize": "^1.1.0",
+      "react-native-animate-number": "^0.1.2",
+      "react-number-easing": "0.0.4",
+      "react-odometerjs": "^2.0.0",
       "react-redux": "^5.0.6",
       "react-router": "^4.2.0",
       "react-router-dom": "^4.2.2",
@@ -68,9 +72,10 @@
       "babel-loader": "^7.1.2",
       "babel-preset-es2015": "^6.24.1",
       "babel-preset-react": "^6.24.1",
-      "nodemon": "^1.12.1",
-      "webpack": "^3.8.1",
       "concurrently": "^3.5.0",
-      "react-scripts": "^1.0.17"
+      "install-peers": "^1.0.2",
+      "nodemon": "^1.12.1",
+      "react-scripts": "^1.0.17",
+      "webpack": "^3.8.1"
    }
 }
